restore typescript client
