+++ exec.php
@@ -93,12 +93,13 @@ require_once(__DIR__ . '/lib/NGXClientGenerator.php');
 $summaryFileName = 'summary.kinf';
 $tmpXmlFileName = tempnam(sys_get_temp_dir(), 'kaltura.generator.');
 
-$options = getopt('hx:r:t:', array(
+$options = getopt('hx:r:t:c:', array(
 	'help',
 	'xml:',
 	'root:',
 	'tests:',
 	'dont-gzip',
+	'custom'
 ));
 
 function showHelpAndExit()
@@ -110,6 +111,7 @@ function showHelpAndExit()
 	echo "\t\t-x, --xml:    \tUse XML path or URL as source XML.\n";
 	echo "\t\t-r, --root:   \tRoot path, default is /opt/kaltura.\n";
 	echo "\t\t-t, --tests:  \tUse different tests configuration, valid values are OVP or OTT, default is OVP.\n";
+	echo "\t\t-c, --custom:  \tSet custom flags (advanced)\n";
 	echo "\t\t--dont-gzip:  \tTar the packages without gzip.\n";
 	
 	exit;
@@ -118,6 +120,7 @@ function showHelpAndExit()
 $schemaXmlPath = null;
 $rootPath = realpath('/opt/kaltura');
 $testsDir = 'ovp';
+$customFlags = null;
 $gzip = true;
 foreach($options as $option => $value)
 {
@@ -137,10 +140,15 @@ foreach($options as $option => $value)
 	{
 		$testsDir = strtolower($value);
 	}
+    elseif($option == 'c' || $option == 'custom')
+    {
+        $customFlags = strtolower($value);
+    }
 	elseif($option == 'dont-gzip')
 	{
 		$gzip = false;
 	}
+
 	array_shift($argv);
 }	 
 
@@ -330,6 +338,12 @@ foreach($config as $name => $item)
 	KalturaLog::info("Generate client library [$name]");
 	$instance->setOutputPath($outputPath, $copyPath);
 	$instance->setTestsPath($testsDir);
+
+	if ($customFlags !== null && method_exists($instance, "setCustomFlags"))
+	{
+	    $instance->setCustomFlags($customFlags);
+	}
+
 	$instance->generate();
 	
 	KalturaLog::info("Saving client library to [$outputPath]");
