+++ sources/ngx/package.json
@@ -37,15 +37,17 @@
     "@angular/compiler": "^5.2.0",
     "@angular/compiler-cli": "^5.2.0",
     "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
     "@types/core-js": "^0.9.37",
     "@types/jest": "^21.1.8",
     "@types/jquery": "^2.0.40",
+    "@types/node": "^8.0.47",
     "core-js": "2.4.1",
     "del": "^2.2.2",
     "gulp": "^3.9.1",
     "gulp-load-plugins": "^1.4.0",
     "jest": "^21.2.1",
-    "jest-preset-angular": "^4.0.1",
+    "jest-preset-angular": "^4.0.2",
     "merge2": "^1.0.2",
     "reflect-metadata": "^0.1.10",
     "rimraf": "^2.5.1",
@@ -58,7 +60,6 @@
     "tslint": "4.3.1",
     "tslint-loader": "3.2.1",
     "typescript": "^2.4.2",
-    "@types/node": "^8.0.47",
     "zone.js": "^0.8.18"
   },
   "engines": {
@@ -66,6 +67,7 @@
     "npm": ">= 5.0.0"
   },
   "jest": {
+    "preset": "jest-preset-angular",
     "setupTestFrameworkScriptFile": "<rootDir>/src/tests/setup-jest.ts",
     "transform": {
       "^.+\\.tsx?$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
