+++ sources/typescript/src/kaltura-client-options.ts
@@ -1,0 +1,6 @@
+export interface KalturaClientOptions {
+    clientTag: string;
+    endpointUrl: string;
+    chunkFileSize?: number;
+    chunkFileDisabled?: boolean;
+}
