--- tests/ovp/ngx/src/api/tests/.gitignore
@@ -1,1 +1,0 @@
-tests-config.ts
