--- tests/ovp/ngx/src/api/tests/tests-config.template.ts
@@ -1,7 +1,0 @@
-export const TestsConfig = {
-    partnerId: "@YOUR_PARTNER_ID@",
-    userName: "testUser",
-    adminSecret: "@YOUR_ADMIN_SECRET@",
-    endpointUrl: "@SERVICE_URL@",
-    clientTag: "ngx-tests"
-};
