+++ tests/ovp/ngx/src/tests/kaltura-file-request.spec.ts
@@ -1,59 +1,68 @@
-import { KalturaClient } from '../kaltura-client.service';
-import { ThumbAssetServeAction } from "../api/types/ThumbAssetServeAction";
-import { BaseEntryListAction } from "../api/types/BaseEntryListAction";
-import { getClient } from './utils';
-import { KalturaClientConfiguration } from '../kaltura-client-configuration.service';
+import {ThumbAssetServeAction} from "../api/types/ThumbAssetServeAction";
+import {getClient} from "./utils";
+import {LoggerSettings, LogLevels} from "../api/kaltura-logger";
+import {KalturaClient} from "../kaltura-client.service";
+import "rxjs/add/observable/throw";
 
 describe("Kaltura File request", () => {
-    // TODO get from utils
-    let kalturaClient: KalturaClient = getClient();
+  let kalturaClient: KalturaClient = null;
 
-    test("thumbasset service > serve action", (done) => {
+  beforeAll(async () => {
+    LoggerSettings.logLevel = LogLevels.error; // suspend warnings
 
-        const thumbRequest = new ThumbAssetServeAction({
-            thumbAssetId: '1_ep9epsxy'
+    return new Promise((resolve => {
+      getClient()
+        .subscribe(client => {
+          kalturaClient = client;
+          resolve(client);
         });
+    }));
+  });
 
-        const config = new KalturaClientConfiguration();
-        config.endpointUrl = 'https://www.kaltura.com';
-        config.clientTag = 'ngxClientTag';
+  afterAll(() => {
+    kalturaClient = null;
+  });
 
-        const predefinedClient = new KalturaClient(null, config);
-        predefinedClient.ks = 'YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs';
+  test("thumbasset service > serve action", (done) => {
 
-        predefinedClient.request(thumbRequest)
-            .subscribe(
-                result => {
-                    expect(result).toBeDefined();
-                    expect(result.url).toBeDefined();
-                    expect(result.url).toBe('https://www.kaltura.com/api_v3/service/thumbasset/action/serve?format=1&apiVersion=3.3.0&thumbAssetId=1_ep9epsxy&ks=YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs&clientTag=ngxClientTag');
-
-                    done();
-                },
-                error => {
-                    const s = '';
-                });
-
+    const thumbRequest = new ThumbAssetServeAction({
+      thumbAssetId: "1_ep9epsxy"
     });
 
-    test("error when sending 'KalturaFileRequest' as part of multi-request", (done) => {
+    kalturaClient.setDefaultRequestOptions({ks: "YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs"});
 
-        const thumbRequest: any = new ThumbAssetServeAction({
-            thumbAssetId: 'thumbAssetId'
+    kalturaClient.request(thumbRequest)
+      .subscribe(
+        result => {
+          expect(result).toBeDefined();
+          expect(result.url).toBeDefined();
+          expect(result.url).toBe("https://www.kaltura.com/api_v3/service/thumbasset/action/serve?clientTag=typescript-tests?apiVersion=3.3.0&ks=YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs&thumbAssetId=1_ep9epsxy&format=1");
+
+          done();
+        },
+        error => {
+          fail(error);
         });
 
+  });
 
-        kalturaClient.multiRequest([thumbRequest])
-            .subscribe(
-                result => {
-                    fail('got response instead of error');
-                },
-                error => {
-                    expect(error).toBeDefined();
-                    expect(error).toBeInstanceOf(Error);
-                    expect(error.message).toBe("multi-request not support requests of type 'KalturaFileRequest'");
-                    done();
-                });
+  test("error when sending 'KalturaFileRequest' as part of multi-request", (done) => {
 
+    const thumbRequest: any = new ThumbAssetServeAction({
+      thumbAssetId: "thumbAssetId"
     });
+
+    kalturaClient.multiRequest([thumbRequest])
+      .subscribe(
+        result => {
+          fail("got response instead of error");
+        },
+        error => {
+          expect(error).toBeDefined();
+          expect(error).toBeInstanceOf(Error);
+          expect(error.message).toBe("multi-request not support requests of type 'KalturaFileRequest', use regular request instead");
+          done();
+        });
+
+  });
 });
