+++ tests/ovp/ngx/src/tests/service-uiconf.spec.ts
@@ -1,25 +1,26 @@
-import { KalturaBrowserHttpClient } from "../kaltura-clients/kaltura-browser-http-client";
-import { UiConfListAction } from "../types/UiConfListAction";
-import { KalturaUiConfListResponse } from "../types/KalturaUiConfListResponse";
-import { KalturaUiConf } from "../types/KalturaUiConf";
-import { KalturaUiConfFilter } from "../types/KalturaUiConfFilter";
-import { KalturaUiConfObjType } from "../types/KalturaUiConfObjType";
-import { UiConfListTemplatesAction } from "../types/UiConfListTemplatesAction";
-import { getClient } from "./utils";
-import { LoggerSettings, LogLevels } from "../kaltura-logger";
+import {UiConfListAction} from "../api/types/UiConfListAction";
+import {KalturaUiConfListResponse} from "../api/types/KalturaUiConfListResponse";
+import {KalturaUiConf} from "../api/types/KalturaUiConf";
+import {KalturaUiConfFilter} from "../api/types/KalturaUiConfFilter";
+import {KalturaUiConfObjType} from "../api/types/KalturaUiConfObjType";
+import {UiConfListTemplatesAction} from "../api/types/UiConfListTemplatesAction";
+import {getClient} from "./utils";
+import {LoggerSettings, LogLevels} from "../api/kaltura-logger";
+import {KalturaClient} from "../kaltura-client.service";
 
 describe(`service "UIConf" tests`, () => {
-  let kalturaClient: KalturaBrowserHttpClient = null;
+  let kalturaClient: KalturaClient = null;
 
   beforeAll(async () => {
     LoggerSettings.logLevel = LogLevels.error; // suspend warnings
 
-    return getClient()
-      .then(client => {
-        kalturaClient = client;
-      }).catch(error => {
-        // can do nothing since jasmine will ignore any exceptions thrown from before all
-      });
+    return new Promise((resolve => {
+      getClient()
+        .subscribe(client => {
+          kalturaClient = client;
+          resolve(client);
+        });
+    }));
   });
 
   afterAll(() => {
@@ -28,7 +29,7 @@ describe(`service "UIConf" tests`, () => {
 
   test("uiconf list", (done) => {
     kalturaClient.request(new UiConfListAction())
-      .then(
+      .subscribe(
         response => {
           expect(response instanceof KalturaUiConfListResponse).toBeTruthy();
           expect(Array.isArray(response.objects)).toBeTruthy();
@@ -46,10 +47,10 @@ describe(`service "UIConf" tests`, () => {
   // TODO [kmc] investigate response
   xtest("get players", (done) => {
     const players = [KalturaUiConfObjType.player, KalturaUiConfObjType.playerV3, KalturaUiConfObjType.playerSl];
-    const filter = new KalturaUiConfFilter({ objTypeIn: players.join(",") });
+    const filter = new KalturaUiConfFilter({objTypeIn: players.join(",")});
 
     kalturaClient.request(new UiConfListAction(filter))
-      .then(
+      .subscribe(
         response => {
           expect(response.objects.every(obj => players.indexOf(Number(obj.objType)) !== -1)).toBeTruthy();
           done();
@@ -69,7 +70,7 @@ describe(`service "UIConf" tests`, () => {
     });
 
     kalturaClient.request(new UiConfListAction(filter))
-      .then(
+      .subscribe(
         response => {
           response.objects.forEach(obj => {
             // TODO [kmc] blocked by previous test-case
@@ -92,7 +93,7 @@ describe(`service "UIConf" tests`, () => {
 
   test("uiconf list templates", (done) => {
     kalturaClient.request(new UiConfListTemplatesAction())
-      .then(
+      .subscribe(
         response => {
           expect(response instanceof KalturaUiConfListResponse).toBeTruthy();
           expect(Array.isArray(response.objects)).toBeTruthy();
