+++ tests/ovp/ngx/src/tests/setup-jest.ts
@@ -1,2 +1,1 @@
-import 'zone.js';
-import 'reflect-metadata';
+import "jest-preset-angular";
