+++ tests/ovp/ngx/src/tests/tests-config.template.ts
@@ -1,0 +1,7 @@
+export const TestsConfig = {
+  partnerId: "@YOUR_PARTNER_ID@",
+  userName: "testUser",
+  adminSecret: "@YOUR_ADMIN_SECRET@",
+  endpointUrl: "@SERVICE_URL@",
+  clientTag: "ngx-tests"
+};
